Got to finally fix the circular import issue, by splitting the configuration code for the Flask app and the database into two files: app_config.py and create_data_db.py
