+++ Gruppe_18/src/main/database/create_data_db.py
@@ -1,0 +1,21 @@
+import os
+
+from sqlalchemy import create_engine
+from sqlalchemy.orm import sessionmaker
+
+from Gruppe_18.src.main.model.models import db
+
+module_path = os.path.dirname(os.path.abspath(__file__))
+database_name = os.path.join(module_path, "../YourGuide.db")
+engine = create_engine(f"sqlite:///{database_name}", echo=True)
+
+session = sessionmaker(bind=engine)()
+db.metadata.create_all(bind=engine)
+
+
+def get_session():
+    engine = create_engine("sqlite:///../YourGuide.db", echo=True)
+
+    Session = sessionmaker(bind=engine)
+
+    return Session()
